# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PathFinder 10k API - AI-powered career guidance for Nigerian students
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudentClassLevel = typeof StudentClassLevel[keyof typeof StudentClassLevel];


export const StudentClassLevel = {
  SSS1: 'SSS1',
  SSS2: 'SSS2',
  SSS3: 'SSS3',
  JAMBite: 'JAMBite',
} as const;

export interface Student {
  id: number;
  clerkId: string;
  fullName: string;
  schoolName: string;
  /** @nullable */
  phoneNumber?: string | null;
  classLevel: StudentClassLevel;
  createdAt: string;
}

export type StudentInputClassLevel = typeof StudentInputClassLevel[keyof typeof StudentInputClassLevel];


export const StudentInputClassLevel = {
  SSS1: 'SSS1',
  SSS2: 'SSS2',
  SSS3: 'SSS3',
  JAMBite: 'JAMBite',
} as const;

export interface StudentInput {
  fullName: string;
  schoolName: string;
  phoneNumber?: string;
  classLevel: StudentInputClassLevel;
}

export type AssessmentAnswers = { [key: string]: unknown };

export interface Assessment {
  id: number;
  studentId: number;
  answers: AssessmentAnswers;
  createdAt: string;
}

export type AssessmentInputAnswers = { [key: string]: unknown };

export interface AssessmentInput {
  answers: AssessmentInputAnswers;
}

export interface Recommendation {
  id: number;
  studentId: number;
  assessmentId: number;
  topCareer: string;
  alternativeCareers: string[];
  jambSubjects: string[];
  /** @nullable */
  aiExplanation?: string | null;
  /** @nullable */
  aiRoadmap?: string | null;
  /** @nullable */
  aiSkills?: string | null;
  /** @nullable */
  aiEncouragement?: string | null;
  isSaved: boolean;
  /** @nullable */
  reasonForRecommendation?: string | null;
  suitableCourses: string[];
  jobRoles: string[];
  requiredSkills: string[];
  createdAt: string;
}

export interface AssessmentResult {
  assessment: Assessment;
  recommendation: Recommendation;
}

export interface CareerPath {
  name: string;
  description: string;
  courses: string[];
  jambSubjects: string[];
  jobRoles: string[];
}

export interface DashboardSummary {
  student?: Student;
  assessmentCount: number;
  savedRecommendationsCount: number;
  latestRecommendation?: Recommendation;
  hasCompletedAssessment: boolean;
}

export interface Conversation {
  id: number;
  studentId: number;
  title: string;
  createdAt: string;
}

export interface ConversationInput {
  title: string;
}

export interface ChatMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatMessageInput {
  content: string;
}

export interface SendMessageResult {
  userMessage: ChatMessage;
  aiMessage: ChatMessage;
}

export type DeleteConversation200 = {
  success: boolean;
};