# lib/api-zod/src/generated/types/conversationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PathFinder 10k API - AI-powered career guidance for Nigerian students
 * OpenAPI spec version: 0.1.0
 */

export interface ConversationInput {
  title: string;
}