# lib/api-zod/src/generated/types/sendMessageResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PathFinder 10k API - AI-powered career guidance for Nigerian students
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface SendMessageResult {
  userMessage: ChatMessage;
  aiMessage: ChatMessage;
}